requirements.txt
python-dotenv
pytest
google-generativeai